Move chat message in js-waku
