+++ src/lib/chat_message/index.ts
@@ -1,6 +1,6 @@
 import { Reader } from 'protobufjs/minimal';
 
-import { ChatMessageProto } from '../proto/chat/v2/chat_message';
+import { ChatMessageProto } from '../../proto/chat/v2/chat_message';
 
 // TODO: Move to waku library?
 export class ChatMessage {
